+++ src/admin/messages/de/cmsadmin.php
@@ -259,4 +259,8 @@ return [
     'menu_group_protocol' => 'Protokoll',
     'menu_group_protocol_model_event_logger' => 'Model Event Log',
     'unable_to_find_item_for_language' => 'Die gesuchte Seite wurde noch nicht übersetzt.',
+
+    // 3.4
+
+    'js_block_attribute_empty' => '«%label%» darf nicht leer sein.',
 ];